Cricket_Data/api/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
python-jose==3.3.0
bcrypt==4.0.1
passlib[bcrypt]
pymysql==1.1.0
cryptography==41.0.7
pydantic[email]==2.5.2
typing-extensions
pandas
